# (Slightly) faster Aarch64 bitpacking.