## What do these changes do?

This PR updates the inline assembly for 64-bit bitpacking on Aarch64.

It uses more vector registers; this makes it possible to pipeline the register load instructions with the next compute instruction (all except the first register load), which wasn't the case previously.

Additionally, the final two `mov` instructions have been replaced with a single `uqshrn` (unsigned saturating shift right and narrow) instruction. This means that the overall compute instruction count (excluding register loads and stores) reduces from 17 to 16.

## How Has This Been Tested?
CI.

## Benchmark Results

I have run benchmarks using the LCE benchmark tool (`num_runs=250`) on a Raspberry Pi.

The results are pretty much within the margin of error, but there seems to be a slight speed up.

| Device | Model         | Threads | Baseline      | PR            | Relative change |
|--------|---------------|---------|---------------|---------------|-----------------|
| RPi 4  | QuickNet      | 1       | 34.58 +- 0.13 | 34.50 +- 0.09 | -0.2%           |
| RPi 4  | QuickNet      | 4       | 17.08 +- 0.06 | 17.12 +- 0.05 | +0.2%           |
| RPi 4  | QuickNetLarge | 1       | 52.45 +- 0.09 | 52.44 +- 0.07 | 0.0%            |
| RPi 4  | QuickNetLarge | 4       | 25.79 +- 0.83 | 25.56 +- 0.06 | -0.9%           |
| RPi 4  | QuickNetXL    | 1       | 92.11 +- 0.22 | 91.83 +- 0.13 | -0.3%           |
| RPi 4  | QuickNetXL    | 4       | 41.69 +- 0.37 | 41.39 +- 0.12 | -0.7%           |

## Related issue number
N/A.